[UI] 폴더 이름 변경 및 샘플 이미지 추가 (UI-#18)
