+++ app/src/main/java/com/texthip/thip/ui/theme/common/form/WarningTextField.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.ui.theme.common.forms
+package com.texthip.thip.ui.theme.common.form
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
@@ -14,7 +14,6 @@ import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
